Fixed problem with long/int conversion.
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Palladio/Incubation/SimuLizar/branches@25817 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ SlidingWindow/org.palladiosimulator.simulizar/src/org/palladiosimulator/simulizar/runtimestate/SimulatedBasicComponentInstance.java
@@ -30,7 +30,7 @@ public class SimulatedBasicComponentInstance extends SimulatedComponentInstance 
         final AssemblyContext myAssCtx = fqID.getAssembyContextPath().get(fqID.getAssembyContextPath().size() - 1);
         for (PassiveResource passiveResource : passiveResources) {
             final long initialCount = StackContext.evaluateStatic(passiveResource.getCapacity_PassiveResource()
-                    .getSpecification(), Integer.class, context.getStack().currentStackFrame());
+                    .getSpecification(), Integer.class, context.getStack().currentStackFrame()).intValue();
             final IPassiveResource simulatedResource = new SimSimpleFairPassiveResource(passiveResource, myAssCtx,
                     getRuntimeState().getModel(), initialCount);
             this.passiveResourcesMap.put(passiveResource.getId(), simulatedResource);
